user_management_2/requirements.txt
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
cryptography==42.0.1
firebase-admin==6.4.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-WTF==1.2.1
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
httplib2==0.22.0
idna==3.6
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.7
pip==22.3.1
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.0.3
pyparsing==3.1.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rsa==4.9
setuptools==65.5.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn[standard]==0.20.0
Werkzeug==2.3.4
WTForms==3.1.2
zipp==3.15.0
gunicorn==20.1.0

